+++ R/generics.R
@@ -286,6 +286,7 @@ print.JANE <- function(x, ...){
 #' @param xlab An optional title for the x axis.
 #' @param ylab An optional title for the y axis.
 #' @param cluster_cols An optional vector of colors for the clusters. Must have a length of at least \eqn{K}.
+#' @param remove_noise_edges (only applicable if \code{JANE} was run with \code{noise_weights = TRUE}) A logical; if \code{TRUE} will remove noise edges based on hard clustering rule of \eqn{\{h | \hat{Z}^{W}_{eh} = max(\hat{Z}^{W}_{e1}, \hat{Z}^{W}_{e2})\}} for \eqn{e = 1,\ldots,|E|}, where \eqn{\hat{Z}^{W}_{e1}} and \eqn{\hat{Z}^{W}_{e2}} are the estimated conditional probabilities that the \eqn{e^{th}} edge is a non-noise and noise edge, respectively (default is \code{FALSE}).
 #' @param ... Unused.
 #' @details
 #'The classification of actors into specific clusters is based on a hard clustering rule of \eqn{\{h | \hat{Z}^{U}_{ih} = max_k \hat{Z}^{U}_{ik}\}}. Additionally, the actor-specific classification uncertainty is derived as 1 - \eqn{max_k \hat{Z}^{U}_{ik}}.
@@ -296,6 +297,8 @@ print.JANE <- function(x, ...){
 #'   \item{\code{termination_rule = 'Q'}: Plots generated are similar to those described in the previous bullet point. However, instead of tracking the change in \eqn{{\hat{Z}^U}} (and potentially \eqn{{\hat{Z}^W}}) over iterations, here the absolute difference in the objective function of the E-step evaluated using parameters from subsequent iterations is tracked. Furthermore, the cumulative average of the absolute change in \eqn{\hat{U}} is no longer tracked.}
 #'  \item{\code{termination_rule \%in\% c('ARI', 'NMI', 'CER')}: Four plots will be presented. Specifically, the top left panel presents a plot of the absolute difference in the cumulative average of the absolute change in the specific \code{termination_rule} employed and \eqn{\hat{U}} across iterations. As previously mentioned, if the EM algorithm converges before the \code{n_its_start_CA}-th iteration then this will be an empty plot. Furthermore, the other three plots present ARI, NMI, and CER values comparing the classifications between subsequent iterations, respectively.}
 #'   }
+#' 
+#' @note If an error interrupts the plotting process, the graphics device may be left in a state where par("new") = TRUE. This can cause subsequent plots to be overlaid. To reset the graphics state, call plot.new() or close and reopen the device with dev.off(); dev.new().
 #'   
 #' @seealso \code{\link[mclust]{surfacePlot}}, \code{\link[mclust]{adjustedRandIndex}}, \code{\link[mclust]{classError}},  \code{\link[aricode]{NMI}}  
 #' 
@@ -354,7 +357,7 @@ print.JANE <- function(x, ...){
 plot.JANE <- function(x, type = "lsnc", true_labels, initial_values = FALSE,
                       zoom = 100, density_type = "contour", rotation_angle = 0,
                       alpha_edge = 0.1, alpha_node = 1, swap_axes = FALSE,
-                      main, xlab, ylab, cluster_cols, ...){
+                      main, xlab, ylab, cluster_cols, remove_noise_edges = FALSE, ...){
   
   if(!inherits(x, "JANE")){
     stop("Object is not of class JANE")
@@ -371,23 +374,70 @@ plot.JANE <- function(x, type = "lsnc", true_labels, initial_values = FALSE,
   if(!type %in% c("lsnc", "trace_plot", "misclassified", "uncertainty")){
     stop("Please provide one of the following for type: 'lsnc', 'trace_plot', 'misclassified', or 'uncertainty'")
   }
-  
+
   opar <- graphics::par(no.readonly = TRUE)
-  on.exit(graphics::par(opar))
+  on.exit({
+    
+    graphics::par(opar)
+    
+    # Warn if the device may be left in an inconsistent state
+    if (graphics::par("new") != opar$new){
+      
+      warning(
+        "Graphics device may be in an inconsistent state (i.e., par(\"new\") = TRUE).\n  To reset it, call `plot.new()` or close and reopen the device with `dev.off(); dev.new().`\n  See Notes section in ?plot.JANE for additional details."
+      )
+      
+    }
+    
+  }, add = TRUE)
   
   trace_plot <- FALSE
   uncertainty <- FALSE
   misclassified <- NULL
   
   if(type == "trace_plot"){
+    
     trace_plot <- TRUE
     plot_data <- x$optimal_res
+    
   } else {
+    
+    temp_A <- x$A
+    
     if(!initial_values){
+      
       plot_data <- x$optimal_res
+      
+      if(remove_noise_edges){
+        if(!x$input_params$noise_weights){
+          warning("Can only remove noise edges if JANE was run with noise_weights = TRUE")
+        } else {
+          
+          Z_W <- plot_data$prob_matrix_W
+          Z_W_labels <- apply(Z_W[, 4:5], 1, which.max)
+          noise <- Z_W[Z_W_labels == 2, 1:2]
+          temp_A[noise] <- 0L
+          
+          if(x$input_params$model %in% c("NDH", "RS")){
+            
+            temp_A[lower.tri(temp_A)] <- t(temp_A)[lower.tri(t(temp_A))]
+            
+          }
+          
+          temp_A <- methods::as(temp_A, "dgCMatrix")
+          
+        }
+      }
+      
     } else {
+      
+      if(remove_noise_edges){
+        message("All edges are assumed to be non-noise when initializing model parameters")
+      }
+      
       plot_data <- x$optimal_starting
     }
+    
   }
   
   plot_data$model <- x$input_params$model
@@ -590,7 +640,7 @@ plot.JANE <- function(x, type = "lsnc", true_labels, initial_values = FALSE,
         cluster_cols
       }
       
-      plot_data(A = x$A,
+      plot_data(A = temp_A,
                 data = plot_data, 
                 misclassified = misclassified,
                 zoom = zoom, 
@@ -651,7 +701,7 @@ plot.JANE <- function(x, type = "lsnc", true_labels, initial_values = FALSE,
           cluster_cols
         }
         
-        plot_data(A = x$A,
+        plot_data(A = temp_A,
                   data = plot_data_temp, 
                   misclassified = misclassified,
                   zoom = zoom,